fixed error

+++ src/main/java/com/Final_Project/Linkedin/repository/ConfirmationTokenRepository.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Repository
-public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Integer> {
+public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Long> {
 
     Optional<ConfirmationToken> findByTokenAndTokenType(String token, TokenType tokenType);
